refactor: book component
