+++ types/book.ts
@@ -1,0 +1,11 @@
+export interface BookRight {
+    dateStart: Date
+    dateEnd?: Date
+    dayStart: number
+    monthStart: number
+    yearStart: number
+    dayEnd: number
+    monthEnd: number
+    yearEnd: number
+    disabledDates: Array<Date>
+}
